ProcessOutResult helper extension functions moved from example to core
