+++ example/src/main/kotlin/com/processout/example/ui/screen/card/CardPaymentViewModel.kt
@@ -4,14 +4,14 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
 import com.processout.example.shared.Constants
-import com.processout.example.shared.getOrNull
-import com.processout.example.shared.onFailure
 import com.processout.example.ui.screen.card.CardPaymentUiState.*
 import com.processout.sdk.api.ProcessOut
 import com.processout.sdk.api.model.request.POCardTokenizationRequest
 import com.processout.sdk.api.model.request.POCreateInvoiceRequest
 import com.processout.sdk.api.repository.POCardsRepository
 import com.processout.sdk.api.service.POInvoicesService
+import com.processout.sdk.core.getOrNull
+import com.processout.sdk.core.onFailure
 import kotlinx.coroutines.async
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
