+++ sdk/src/main/kotlin/com/processout/sdk/core/ProcessOutResult.kt
@@ -23,6 +23,55 @@ sealed class ProcessOutResult<out T : Any> {
     ) : ProcessOutResult<Nothing>()
 }
 
+inline fun <T : Any> ProcessOutResult<T>.onSuccess(
+    action: (value: T) -> Unit
+): ProcessOutResult<T> {
+    if (this is Success) {
+        action(value)
+    }
+    return this
+}
+
+inline fun <T : Any> ProcessOutResult<T>.onFailure(
+    action: (failure: Failure) -> Unit
+): ProcessOutResult<T> {
+    if (this is Failure) {
+        action(this)
+    }
+    return this
+}
+
+inline fun <T : Any, R> ProcessOutResult<T>.fold(
+    onSuccess: (value: T) -> R,
+    onFailure: (failure: Failure) -> R
+): R = when (this) {
+    is Success -> onSuccess(value)
+    is Failure -> onFailure(this)
+}
+
+fun <T : Any> ProcessOutResult<T>.getOrNull(): T? =
+    when (this) {
+        is Success -> value
+        is Failure -> null
+    }
+
+inline fun <T : Any, R : Any> ProcessOutResult<T>.map(
+    transform: (T) -> R
+): ProcessOutResult<R> = when (this) {
+    is Success -> Success(transform(value))
+    is Failure -> this.copy()
+}
+
+fun <T : Any> ProcessOutResult<T>.copy(): ProcessOutResult<T> =
+    when (this) {
+        is Success -> this.copy()
+        is Failure -> this.copy()
+    }
+
+@Deprecated(
+    message = "Use replacement function.",
+    replaceWith = ReplaceWith("onSuccess(action)")
+)
 inline fun <T : Any> ProcessOutResult<T>.handleSuccess(
     block: (value: T) -> Unit
 ) {
@@ -31,6 +80,10 @@ inline fun <T : Any> ProcessOutResult<T>.handleSuccess(
     }
 }
 
+@Deprecated(
+    message = "Use replacement function.",
+    replaceWith = ReplaceWith("onFailure(action)")
+)
 inline fun <T : Any> ProcessOutResult<T>.handleFailure(
     block: (
         code: POFailure.Code,
@@ -42,17 +95,4 @@ inline fun <T : Any> ProcessOutResult<T>.handleFailure(
     if (this is Failure) {
         block(code, message, invalidFields, cause)
     }
-}
-
-fun <T : Any> ProcessOutResult<T>.copy(): ProcessOutResult<T> =
-    when (this) {
-        is Success -> this.copy()
-        is Failure -> this.copy()
-    }
-
-inline fun <T : Any, R : Any> ProcessOutResult<T>.map(
-    transform: (T) -> R
-): ProcessOutResult<R> = when (this) {
-    is Success -> Success(transform(value))
-    is Failure -> this.copy()
 }
